+++ R/env.R
@@ -111,13 +111,13 @@
 #' env$a
 env <- function(...) {
   env <- new.env(parent = caller_env())
-  env_bind(env, !!! dots_list(...))
+  env_bind(.env = env, ...)
 }
 #' @rdname env
 #' @export
 child_env <- function(parent, ...) {
   env <- new.env(parent = as_env(parent))
-  env_bind(env, ...)
+  env_bind(.env = env, ...)
 }
 
 #' Coerce to an environment