+++ NEWS.md
@@ -1,6 +1,10 @@
 
 # rlang 0.1.9000
 
+* `env()` and `env_child()` can now get arguments starting with `.`.
+  Prior to this fix, these arguments were partial-matching on
+  `env_bind()`'s `.env` argument.
+
 
 # rlang 0.1